#192481 virtual items and game keys price block refactoring
